676: refactor some parts of the campaign service to reuse getFromApi
