+++ lib/app/services/gruene_api_door_service.dart
@@ -1,13 +1,13 @@
 import 'package:gruene_app/app/services/converters.dart';
 import 'package:gruene_app/app/services/enums.dart';
-import 'package:gruene_app/app/services/gruene_api_campaigns_service.dart';
+import 'package:gruene_app/app/services/gruene_api_campaigns_base_service.dart';
 import 'package:gruene_app/features/campaigns/models/doors/door_create_model.dart';
 import 'package:gruene_app/features/campaigns/models/doors/door_detail_model.dart';
 import 'package:gruene_app/features/campaigns/models/doors/door_update_model.dart';
 import 'package:gruene_app/features/campaigns/models/marker_item_model.dart';
 import 'package:gruene_app/swagger_generated_code/gruene_api.swagger.dart';
 
-class GrueneApiDoorService extends GrueneApiCampaignsService {
+class GrueneApiDoorService extends GrueneApiCampaignsPoiBaseService {
   GrueneApiDoorService() : super(poiType: PoiServiceType.door);
 
   Future<MarkerItemModel> createNewDoor(DoorCreateModel newDoor) async => getFromApi(
